Create StringBuilderDemo.java
The below code is used to understand the difference between String and StringBuilder objects. When we try to concatenate two strings using string operations a new object is created without changing the old one. In StringBuilder existing object is modified. In the below program this can be illustrated by comparing Hash Code for String object after every concat operation. Fill the missing code in the below program and observe the output.